// web/src/api/generated/model/queueTab.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */

export type QueueTab = typeof QueueTab[keyof typeof QueueTab];


export const QueueTab = {
  to_review: 'to_review',
  needs_checking: 'needs_checking',
  waiting: 'waiting',
  decided: 'decided',
} as const;